move values in wasm backend to 64 bit. it still has truncation problems for huge values, but given that wasm64 is a thing, this is probably the right direction

+++ rolandc/src/backend/wasm.rs
@@ -52,12 +52,12 @@ enum ContainingSyntax {
 struct GenerationContext<'a> {
    active_fcn: wasm_encoder::Function,
    type_manager: TypeManager<'a>,
-   literal_offsets: HashMap<StrId, (u32, u32)>,
+   literal_offsets: HashMap<StrId, (u64, u64)>,
    global_info: &'a IndexMap<VariableId, GlobalInfo>,
-   static_addresses: HashMap<VariableId, u32>,
-   stack_offsets_mem: HashMap<usize, u32>,
+   static_addresses: HashMap<VariableId, u64>,
+   stack_offsets_mem: HashMap<usize, u64>,
    user_defined_types: &'a UserDefinedTypeInfo,
-   sum_sizeof_locals_mem: u32,
+   sum_sizeof_locals_mem: u64,
    proc_name_table: &'a HashMap<StrId, ProcedureId>,
    procedure_to_table_index: IndexSet<ProcedureId>,
    procedure_indices: IndexSet<ProcedureId>,
@@ -253,7 +253,7 @@ pub fn emit_wasm(
 
    // the base memory offset varies per platform;
    // on wasm-4/microw8, we don't own all of the memory!
-   let mut offset: u32 = match config.target {
+   let mut offset: u64 = match config.target {
       Target::Generic | Target::QbeFreestanding | Target::QbeHost => unreachable!(),
       Target::Wasi => 0x1, // Don't put anything at 0 to allow null to be a sentinel value
       Target::Wasm4 => 0x19a0,
@@ -267,8 +267,7 @@ pub fn emit_wasm(
          &ConstExpr::i32_const(offset as i32),
          str_value.as_bytes().iter().copied(),
       );
-      //TODO: and here truncation
-      let s_len = str_value.len() as u32;
+      let s_len = str_value.len() as u64;
       generation_context.literal_offsets.insert(*s, (offset, s_len));
       offset += s_len;
    }
@@ -289,7 +288,7 @@ pub fn emit_wasm(
          1
       };
 
-      offset = aligned_address(u64::from(offset), strictest_alignment) as u32;
+      offset = aligned_address(offset, strictest_alignment);
    }
    for (static_var, static_details) in program.non_stack_var_info.iter() {
       debug_assert_ne!(static_details.kind, StorageKind::Const);
@@ -304,7 +303,7 @@ pub fn emit_wasm(
          &static_details.expr_type.e_type,
          generation_context.user_defined_types,
          BaseTarget::Wasm,
-      ) as u32;
+      );
    }
 
    let mut buf = vec![];
@@ -324,7 +323,7 @@ pub fn emit_wasm(
    }
 
    // keep stack aligned
-   offset = aligned_address(u64::from(offset), 8) as u32;
+   offset = aligned_address(offset, 8);
 
    let (global_section, global_names) = {
       let mut globals = GlobalSection::new();
@@ -453,18 +452,18 @@ pub fn emit_wasm(
 
       generation_context.sum_sizeof_locals_mem = 0;
 
-      let mut mem_info: IndexMap<usize, (u32, u32)> = regalloc_result.procedure_stack_slots[proc_id]
+      let mut mem_info: IndexMap<usize, (u64, u64)> = regalloc_result.procedure_stack_slots[proc_id]
          .iter()
          .enumerate()
-         .map(|(i, x)| (i, (x.1 as u32, x.0 as u32)))
+         .map(|(i, x)| (i, (x.1, x.0)))
          .collect();
 
-      mem_info.sort_by(|_k_1_, v_1, _k_2_, v_2| compare_alignment(u64::from(v_1.0), u64::from(v_1.1), u64::from(v_2.0), u64::from(v_2.1)));
+      mem_info.sort_by(|_k_1_, v_1, _k_2_, v_2| compare_alignment(v_1.0, v_1.1, v_2.0, v_2.1));
 
       for local in mem_info.iter() {
          // last element could have been a struct, and so we need to pad
          generation_context.sum_sizeof_locals_mem =
-            aligned_address(u64::from(generation_context.sum_sizeof_locals_mem), u64::from(local.1.0)) as u32;
+            aligned_address(generation_context.sum_sizeof_locals_mem, local.1.0);
          generation_context
             .stack_offsets_mem
             .insert(*local.0, generation_context.sum_sizeof_locals_mem);
@@ -1708,7 +1707,7 @@ fn get_stack_address_of_local(id: VariableId, generation_context: &mut Generatio
    let Some(VarSlot::Stack(s)) = generation_context.var_to_slot.get(&id) else {
       return false;
    };
-   let offset = aligned_address(u64::from(generation_context.sum_sizeof_locals_mem), 8) as u32
+   let offset = aligned_address(generation_context.sum_sizeof_locals_mem, 8)
       - generation_context
          .stack_offsets_mem
          .get(&(*s as usize))
@@ -1846,7 +1845,7 @@ fn adjust_stack(generation_context: &mut GenerationContext, instr: &Instruction)
 
    generation_context.active_fcn.instruction(&Instruction::GlobalGet(SP));
    // ensure that each stack frame is strictly aligned so that internal stack frame alignment is preserved
-   let adjust_value = aligned_address(u64::from(generation_context.sum_sizeof_locals_mem), 8);
+   let adjust_value = aligned_address(generation_context.sum_sizeof_locals_mem, 8);
    generation_context
       .active_fcn
       .instruction(&Instruction::I32Const(adjust_value as i32));
